feat: add backend for roles and permissions

+++ internal/app/app.go
@@ -28,22 +28,22 @@ type AppLayer interface {
 	CreateAsset(user *models.UserInternal, fileName, contentType, data string) (*models.AssetInternal, error)
 	CreateLocation(user *models.UserInternal, name, mainImageName, individualImageName, backgroundImagePath, color, address, startTime, endTime, description string) (*models.LocationInternal, error)
 	CreateRedirect(user *models.UserInternal, fromPath, toUrl string, startsOn, stopsOn int64) (*models.RedirectInternal, error)
-	DeleteAsset(id uint64) error
-	DeleteLocation(id uint64) error
-	DeleteRedirect(id uint64) error
+	DeleteAsset(id uint) error
+	DeleteLocation(id uint) error
+	DeleteRedirect(id uint) error
 	FindAsset(fileName string) (string, error)
 	FindRedirect(path string) (*models.RedirectInternal, error)
 	GetAllAssets() (*[]models.AssetInternal, error)
 	GetAllLocations() (*[]models.LocationInternal, error)
 	GetAllRedirects() (*[]models.RedirectInternal, error)
-	GetAsset(id uint64) (*models.AssetInternal, error)
-	GetLocation(id uint64) (*models.LocationInternal, error)
-	GetRedirect(id uint64) (*models.RedirectInternal, error)
+	GetAsset(id uint) (*models.AssetInternal, error)
+	GetLocation(id uint) (*models.LocationInternal, error)
+	GetRedirect(id uint) (*models.RedirectInternal, error)
 	GetUser(userId uint) (*models.UserInternal, error)
-	UpdateAsset(user *models.UserInternal, id uint64, fileName, contentType, data string) (*models.AssetInternal, error)
+	UpdateAsset(user *models.UserInternal, id uint, fileName, contentType, data string) (*models.AssetInternal, error)
 	UpdatePassword(user *models.UserInternal, password string) error
-	UpdateLocation(user *models.UserInternal, id uint64, name, mainImageName, individualImageName, backgroundImagePath, color, address, startTime, endTime, description string) (*models.LocationInternal, error)
-	UpdateRedirect(user *models.UserInternal, id uint64, fromPath, toUrl string, startsOn, stopsOn int64) (*models.RedirectInternal, error)
+	UpdateLocation(user *models.UserInternal, id uint, name, mainImageName, individualImageName, backgroundImagePath, color, address, startTime, endTime, description string) (*models.LocationInternal, error)
+	UpdateRedirect(user *models.UserInternal, id uint, fromPath, toUrl string, startsOn, stopsOn int64) (*models.RedirectInternal, error)
 	ValidatePassword(user *models.UserInternal, password string) error
 }
